Initial migration of project from SourceForge.
This project was formerly hosted at

https://sourceforge.net/projects/qform/

The code in this commit was released as part of the last formal release
of QueryForm, version 1. There were a few other features added to the
code after that by other developers, but those features were never
released. The code for those features is not included in this commit
because I received no response from the authors when I contacted them to
let them know about the migration. All of the code in this commit was
written by the owner of this project, David Glasser.